+++ src/RenderSystem.GL4/VertexBufferGL4.hpp
@@ -26,13 +26,13 @@ public:
 	VertexBufferGL4() = delete;
 	~VertexBufferGL4() override;
 
-	VertexBufferGL4(std::uint32_t sizeInBytes, BufferUsage bufferUsage);
+	VertexBufferGL4(std::size_t sizeInBytes, BufferUsage bufferUsage);
 
-	VertexBufferGL4(void const* vertices, std::uint32_t sizeInBytes, BufferUsage bufferUsage);
+	VertexBufferGL4(void const* vertices, std::size_t sizeInBytes, BufferUsage bufferUsage);
 
 	///@copydoc NativeVertexBuffer
-	void SetData(std::uint32_t offsetInBytes, void const* source,
-		std::uint32_t sizeInBytes) override;
+	void SetData(std::size_t offsetInBytes, void const* source,
+		std::size_t sizeInBytes) override;
 
 	void BindBuffer();
