+++ src/RenderSystem/NativeGraphicsDevice.hpp
@@ -62,12 +62,12 @@ public:
 	///@~Japanese
 	/// @brief 頂点バッファを作成します。
 	virtual std::unique_ptr<NativeVertexBuffer>
-	CreateVertexBuffer(std::uint32_t sizeInBytes, BufferUsage bufferUsage) = 0;
+	CreateVertexBuffer(std::size_t sizeInBytes, BufferUsage bufferUsage) = 0;
 
 	///@~Japanese
 	/// @brief 頂点バッファを作成します。
 	virtual std::unique_ptr<NativeVertexBuffer>
-	CreateVertexBuffer(void const* vertices, std::uint32_t sizeInBytes,
+	CreateVertexBuffer(void const* vertices, std::size_t sizeInBytes,
 		BufferUsage bufferUsage) = 0;
 
 	///@~Japanese
