+++ src/RenderSystem/NativeVertexBuffer.hpp
@@ -9,7 +9,7 @@
 #endif
 
 #include "Pomdog/Graphics/detail/ForwardDeclarations.hpp"
-#include <cstdint>
+#include <cstddef>
 
 namespace Pomdog {
 namespace Detail {
@@ -29,8 +29,8 @@ public:
 	/// @param offsetInBytes バッファの先頭から出力先までのバイト単位のオフセット
 	/// @param source ソースバッファ
 	/// @param sizeInBytes 頂点データのバイト単位のサイズ
-	virtual void SetData(std::uint32_t offsetInBytes,
-		void const* source, std::uint32_t sizeInBytes) = 0;
+	virtual void SetData(std::size_t offsetInBytes,
+		void const* source, std::size_t sizeInBytes) = 0;
 };
 
 }// namespace RenderSystem
